hide interruptions handled by emulate_ptrace
a user's wait() returned at the first interruption of the process even if emulate_ptrace was already handling it. Now if the interruption isn't caused by a breakpoint or the user we let waitpid decide if the user should be informed. This may cause problems if waitpid is never reached, but it is better that having the user's script run while he are emulating the process