warn problems split_on_fork with interrupt
The current implementation of split_on_fork relies on the callback on the new_child event. This will be executed so late that the parent is usually already waiting for the user input when we can finally stop it, making the interrupt option useless.

The only safe way currently is for the user to set a breakpoint on the fork function. I would like to have something that always works based on the syscall though.