cache libc
p.libc takes 0.5s every time to check the maps file! Once we find the libc instead we just save it.

This is still a slow when using symbols before the libc is loaded because we search for it every time, but we can look into events.new_objfile to update a dictionary with all the libraries instead of creating it every time from maps. (Removing the logs should also reduce the usage of symbols by reverse_lookup)