cache instruction pointer
GDB can only access registers with selected_frame().read_register or parse_and_eval, which takes around 0.05s.

We may want to cache only the instruction pointer and only use events.cont to clear it, but it requires to check if the register we are accessing is an instruction pointer.
We could improve it by clearing only the register that was changed by the user, but requires to know which regnum is which register.