move stop handler thread outside event callback
pwntools is already using event.stop to handle the gdb.wait function

Adding an additional callback to handle our interruptions makes it twice as slow (2ms -> 4ms). Removing it is not a big gain, and we could obtain the same results in a simpler way by deleted pwntools' callback, but taking the Thread outised gdb allow us to keep the context and prevents an obscure bug when running as root and trying to create a new debugger from a callback. Although it's more crucial in split_on_fork