Set context if forgot by user
We absolutely need the context to be set in the thread the script is running. It may cause problems it some users need to change the context through the script, but they should also know about context.local in that case.

(And the old restore_arch was creating worse problems in that case)