support qemu with GEF
On ubuntu 24.04 and with arm 64bit as I'm testing now it doesn't seem to be a problem, but /proc/pid/maps wasn't present when we developed the support for qemu, so we are using vmmap from gdb. pwndbg and gef use different representations though and have to be parsed differently.