ltrace
Putting the breakpoint on the plt means that we are tracing only the calls from the binary to the library. This should be the intended usage so I don't see a reason to trace all the function in the library.
Furthermore it is safe to call finish because the calls to the plt should be real calls (not jump) and the library will never go back to the plt before returning