LIFO callback resolution
It feels more natural to set ltrace first before adding other breakpoints. In that case the callback on the plt would be executed after the function has returned, so the user callback wouldn't be executed at the expected address anymore. I feel more natural executing them in a LIFO order, but if it's not we could just had a warning when setting a callback on the plt.