prioritize target to find elf over context.binary
context.binary is usually set once to start debugging with the correct architecture and then may not change if we are debugging another app in the same architecture, so if it's needed to know precisely which binary we are using it is as simple to pass it to the debugger object than to set the context again. For this reason changing context.binary will not be mandatory to debug two different programs.