hotfix prevent from_entry when attaching
We do not currently check that we call until_entry only when the program is in the loader. This is a problem when attaching to a running process.

When splitting after a fork we can set ourself from_start = False to tell Debugger we are not debugging from the beginning of the program, but I would prefer soon to have a check in which file we are. It will probably require to update the ELF objects to allow x in ELF() to check that we are in the range of min and max address