Hotfix using unsigned 32 bit registers
We are always extracting the registers as unsigned ints. From Ubuntu 24.04 we can use .bytes to get the exact value, but in Ubuntu 22.04 we have to use parse_and_eval which may return negative values. In this case if the register is 32 bit and the context is 64 bit, doing % 2**context.bits is wrong.

Can we define register objects to make sure we know the size ? This will help caching properly the registers, but may not be indispensable since the main focus of the cache is to keep the instruction pointer and this is still working as it is set as soon as we read it the first time.