Introduce EXE class
We want an extension of the ELF class so that we are able to check if an address is inside the library or not.
This will allow us to support also PE binaries on day hopefully.

The way to find the address of a binary now changes as we can rely on libs() to find both base address and ending address.