hotfix relative addresses must be small
I still don't know if we should make sure that the address is only of the size of the code section of the binary or that it would end up inside the pages allocated to it. I feel like we want the first one most of the time, but the second is safer as long as it doesn't create a range too large.