[Hotfix] self.libs support non debugging
We expect libs to return the local path to all libraries used by the program. When debugging the program we can retrieve them ourselves, but otherwise we may still rely on it to find a local copy of the libraries running remotely. In this case we fall back to the implementation in pwntools that spwns a local process to find them.

This is not guaranteed to work with QEMU. In that case we would need to debug the program ourselves. This is slow though, so I would really want to limit it as much as possible and check how pwntools runs an headless version for testing.