[Feature] Allow argument ASLR to overwrite eventual aslr=False in script
The idea is that for simplicity you may want to disable ASLR when debugging your program so that you can recognize the different addresses, but you still want to test quickly with ASLR on that the payload still works. Instead of having to edit the code I prefer to support a new argument that forces ASLR on.