[Fix] Support different context in same script
context is set for the whole script, but we may need to debug at the same time a program on amd64 and on riscv (for example debugging qemu and the emulated program at the same time), so the context must be constrained to the Debugger object. We already introduced __context_params to keep the Threads running with the right context, but a lot of calls to context.native where made outside of threads.

We could try to optimize when we set the context or directly access __context_params, but I'll look into it only if this slows down too much the debugging