[Fix] Handle continue() before syscall
We used to call a step before calling continue because we want to make sure we digest any signal left. This may not be needed now with stop_reason, but would require work before we can remove it.

The main case where it was a problem was after an interrupt, because the most likely place to stop the process is while it's waiting to read user input.

For now I introduce a timeout in case that step hits a syscall such as read that would not return